Feat #2 : S3 이미지 업로드 Controller 작성하기

+++ src/main/java/project/backend/domain/ticket/dto/TicketPatchRequestDto.java
@@ -15,7 +15,7 @@ public class TicketPatchRequestDto {
     public String imageUrl;
     public String ticketImageUrl;
     public LocalDateTime ticketDate;
-    public Integer rating;
+    public Float rating;
     public String memo;
     public String seat;
     public String location;
